♻️refactor(column) StatusMessage.java 패키지 이동
