+++ src/main/java/com/sparta/bitbucket/column/controller/ColumnController.java
@@ -1,6 +1,6 @@
 package com.sparta.bitbucket.column.controller;
 
-import static com.sparta.bitbucket.column.dto.StatusMessage.*;
+import static com.sparta.bitbucket.column.entity.StatusMessage.*;
 
 import java.util.List;
